backend/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.42.2
dnspython==2.7.0
flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
idna==3.10
importlib-metadata==8.6.1
itsdangerous==2.2.0
jinja2==3.1.5
MarkupSafe==3.0.2
marshmallow==3.26.1
packaging==24.2
PyJWT==2.10.1
pymongo==4.11
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
urllib3==2.3.0
werkzeug==3.1.3
zipp==3.21.0
